tests/wpt/tests/html/semantics/interactive-elements/the-dialog-element/dialog-requestclose-3.html

<!doctype html>
<meta charset="utf-8">
<link rel=help href="https://html.spec.whatwg.org/multipage/interactive-elements.html#dom-dialog-request-close">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="author" href="mailto:lwarlow@igalia.com" />

<script>
  test(() => {
      const doc = document.implementation.createHTMLDocument("");
      const dialog = doc.createElement("dialog");
      doc.body.append(dialog);
      dialog.setAttribute("open", "");
      dialog.requestClose();
      assert_true(dialog.open);
  }, "requestClose() should not close the dialog when is in inactive document");
</script>
