LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-requestclose-3-expected.txt


PASS requestClose() should not close the dialog when is in inactive document
